Polish Styling:
Add color and content to define the styling of
the entire app.

The Form styling remains rather simple, however,
cards are a little more worked by adding a toggle
button.

WARNING: The toggle button behavior does not work
as expected yet. When clicking on any card, it
affects the original card only.

Add a footer on the side bar with a link to the
repository