Add function:
Define the displayBook function which handles
the DOM behavior to display the array items.

This isn't the final version as it has un-scalable
features like displaying only the first item of
the array when invoked.

Therefore, this commit purpose is to leave record
of the DOM events behavior, creating a div and
appending that div into the DOM.