Finish Function & Fix bug:
Finish adding the content for displayBook()
function, which was defined in previous commit.

On this commit, I added the rest of the content
besides title.

Bugfix:

In previous commits, there was an issue with cards
toggle checkbox only working for the original card

This is because all of the inputs shared the exact
same id.

On this commit, we add the counter variable which
helps create a unique id for each input.

Another option could have been to use:

    crypto.randomUUID();

It was avoided to avoid making the DOM to lengthy