Feat: Create event listener for delete button
This commit defines the logic to delete both the
book ux card and remove the book object from the
library array when the delete button is pressed.

The ux logic is simple, we use .removeChild() to
remove the corresponding div container from the
main content div. We target "newBookDiv" which
exists within the scope of the displayBook()
function, knowing which card it
should target...afaik.

The array logic is not as simple, since you can't
just target the array item based on a value. An
index position is needed for splice to know what
to remove.

Therefore, we created "targetIndex" which is a
variable limited to the scope of the event listener
that uses .findIndex() to get the exact index of
the corresponding newBook.id

This is where the use of crypto.randomUUID() pays
off. Any other identifiable value can be
duplicated, except that one, therefore knowing
its exact target. If we used newBook.author, the
behavior would be unexpected for possible homonyms

targetIndex gets the index of the target, then,
splice uses it to know which index to to remove
and a 1 is passed as the second arg so that it only
removes one item at a time.

A loop was used to create 7 instances of the same
book, differentiated only by their unique id's.

console.log is present to show the original 7 books
arrangement and how it gets coherently updated each
time the delete button is pressed.